[Rust SDK][refactor](ast_module): Renamed module as ast_module.
